rename atomized chain
